Moves sprite classes into new subpackage
