+++ src/pyrite/sprite/__init__.py
@@ -1,0 +1,2 @@
+from .sprite import Sprite  # noqa:F401
+from .spritesheet import SpriteMap, SpriteSheet  # noqa:F401
