+++ src/pyrite/sprite/spritesheet.py
